Add Kafka data streams monitoring implementation (#3192)
* Add support for Data Streams Monitoring to Kafka

If enabled, checkpoints and injects the pathway context into Kafka headers when producing Kafka messages.
If enabled, extracts, merges (for fan-in situations), and sets a checkpoint when consuming.

As far as I can tell, in the "default" scenarios (`KafkaCreateConsumerScopeEnabled` is true and processing in a loop; `KafkaCreateConsumerScopeEnabled` is false and consuming multiple before processing) we will never need to "merge" contexts. Instead, we may have to expose the "merge" functions for users running a fan-in scenario. Need to think about the API to expose for that, and somehow incorporate into the ParentContext extraction helper

* Add Data Streams Monitoring integration test and snapshot

Reused a lot of the code from Samples.Kafka for creating consumers and producing messages, so extracted a base class to help with that.

The snapshots are rough but the best I can do.

* Don't create a new array every time we set a "producer" checkpoint in Kafka

* PR feedback

* Try make topic create/delete code more robust

* Add log of number of values serialized for debugging

* Revert "Add log of number of values serialized for debugging"

This reverts commit 1def339af3d2e41362e5ff59edb42e9e45b51ae7.