Throughput for blocking (#3277)
## Summary of changes

Adds a throughput test for blocking.

## Implementation details

Blocking is now based on the WAF rules, to be able to block from the perf tests we needed a rule that would enable blocking from just the url path. This meant:
* enabling the throughput tests to use a custom file
* changing the snapshots to take into account the simplification of the test blocking rule.

It also fixes the use of the profiler in the adhoc throughput environment that had been broken by the refactoring of the home folder. 

## Test coverage

It is a test.