Enable agent-based telemetry by default (#2800)
* minor: Downgrade telemetry send failures to info

So as not to worry users. It's not critical for them, so I don't think we should be worrying them in the logs.

* Extract LogsApiTests as a helper for use in other tests

* Enable telemetry by default

* Add support for sending telemetry using two transports

- Renamed the TelemetryCircuitBreaker to TelemetryTransportManager as it handles the transport failover as well
- For the agent transport, if the request is not successful, attempt to read the agent version, and if it's not high enough, then don't bother with any retries
- Currently, we only check the agent on an error - we _could_ capture it on success too, but can't see a scenario where that would be useful
- We currently extract the version separately from the trace API. I looked at sharing that across, but I don't think there's a lot of value, and it adds quite a lot of complexity.

* Add additional logging

* Add configuration key for enabling (disabling) agent proxy telemetry

This is primarily useful in testing scenarios, but it also allows us to skip trying to send to the agent when we're in scenarios we know we don't have an agent, or where we _probably_ don't have an agent. The latter is the case for CIApp when in agentless mode, and may well become more prevalent in that future.

* Removed agent version check from AgentlessTelemetryTransport

The agent doesn't send the AgentVersion header if Telemetry is not supported (404) which makes this check a bit pointless (we will only receive the header if Telemetry _is_ supported)

We _could_ look at reading the agent version from the tracer `API`, but it all seems a little pointless

Removed TelemetryPushResult.FatalErrorDontRetry code paths as they will never be hit now

* Remove unused parameter from tests

* If no valid telemetry transports, disable telemetry

Not _explicitly_ disabling telemetry, just using `NullTelemetryController`

* and some forgotten changes

* Revert changes to LogsApiTests

* Update configuration key comment

Co-authored-by: Pierre Bonet <pierre.bonet@datadoghq.com>