[ASM] Remove warnings and fix flake (#3338)
## Summary of changes

We are seeing flake in the ASM blocking tests, where the status code and resource name disappeared from the span:

![framework flake](https://user-images.githubusercontent.com/51381/195104478-2c131e57-f8f0-46a9-bd0b-01a0fb1f0758.png)


## Reason for change

Flake is bad.

## Implementation details

Before this change we dispose systematically disposed a scope / span in the `TracingHttpModule.OnBeginRequest` in the case of an exception, the logic being we won't be able to retrieve it from the http context `TracingHttpModule.OnEndRequest`.

However, this is not always true, if the exception occurs late `TracingHttpModule.OnBeginRequest`, then the span will be in the http context and we'll attempt to update the `TracingHttpModule.OnEndRequest`. This creates a race between data being written to the span object and the span being written to the `AgentWriter`.

In the case where we blocked the request from the call to the WAF in `TracingHttpModule.OnBeginRequest`, we were also calling the WAF again in `TracingHttpModule.OnEndRequest`, which seems unnecessary.