Make cgroup ContainerMetadata parsing more lenient
We were failing to extract the containerID from cgroup in an EKS1.23 nodegroup with Amazon Linux, due to the regex we use for extraction.

This relaxes the Regex somewhat. Instead of matching the start of the line, we only try to match the end of the line, and match either the task, uuid, or containerID regex.