Create custom image for building in Gitlab
Previously, we were using the Windows agent runners, but due to differing dependencies, that's no longer possible.

Instead, this defines a custom Docker image which we build externally and push to dockerhub.

Only installs the minimal required dependencies. Many of the scripts are taken from the agent build repo (and tweaked)