Add TelemetryLogsSink that sends logs to telemetry back-end
This is a little tricky, as it can only start sending _after_ telemetry is initialized, and it may need to replace _where_ it's sending to.

Additionally, it can't log anything (otherwise can get into circular dependency issues

Finally, unlike most telemetry, we have to be more aggressive with dropping after failed attempts, so taken a simpler approach to retries/cycling of transports.