Adds resource attributes to the Datadog spans generated by OpenTelemetry spans. This also fixes an issue so that spans generated from other OpenTelemetry Tracers will now have the same service name as the service attribute, matching the .NET OTLP exporter behavior.
This is achieved by adding automatic instrumentation to the method OpenTelemetry.Trace.TracerProviderBuilderExtensions.Build that adds a new Processor to the user-created TracerProviderBuilder. This is needed because the resource attributes are applied per TracerProvider instance and the best way to access the resources is from a processor which is able to access the original configuration. This processor responds to the OnStart event, gets the Datadog span that corresponds to the newly created Activity object, and then updates the Datadog span with the resource attributes.

This is tested by testing the default service resource attributes that already exist in the Samples.OpenTelemetrySdk application:
- service.instance.id
- service.name
- service.version