Skip GraphQL 2 tests on .NET Framework on x64
This test uses ASP.NET Core running on .NET Framework, which uses the native libuv library. By default, if you don't compile for a specific architecture, the SDK assumes you're going to run on x86 (default for IIS) so copies the x86 native lib. As we now compile for AnyCPU, that means we now have the wrong architecture and would crash.

Not worth the effort in build compexity to resolve, so just skipping the test in this one case for now.