Fix x86 debugger tests by adjusting strong-naming
As we are modifying the sample exe, it can't be strong-named, otherwise the signature is invalid.

Currently Samples.Probes.exe is strong-named, but it doesn't need to be; only its _dependencies_ are referenced by the integration project (which is strong named, because Datadog.Trace is strong named and we have InternalsVisibleTo).

This commit fixes the issue by moving the strong-naming to the dependencies only, meaning we can safely update Samples.Probes.exe