Rework build to always build for AnyCPU instead of x64/x86 (#3632)
* Remove <Platforms> for x64/x86 from all projects

* Stop building for x64/x86, and use AnyCPU instead

* Remove CreatePlatformlessSymlinksTarget

As we're always going to build any_cpu, so this shouldn't be necessary

* Update tests to use provided TargetPlatform env var

* Fix running x86 tests on Windows .NET Framework by forcing the CorFlags on the exe before executing

This feels pretty hacky, but avoids a bunch of other build issues by not specifying x64/x86 etc

* Fix security ASP.NET projects incorrectly setting PlatformTarget in AnyCpu configuration

* Fix regression tests by reusing ProfilerHelper

* Skip GraphQL 2 tests on .NET Framework on x64

This test uses ASP.NET Core running on .NET Framework, which uses the native libuv library. By default, if you don't compile for a specific architecture, the SDK assumes you're going to run on x86 (default for IIS) so copies the x86 native lib. As we now compile for AnyCPU, that means we now have the wrong architecture and would crash.

Not worth the effort in build compexity to resolve, so just skipping the test in this one case for now.

* Fix x86 debugger tests by adjusting strong-naming

As we are modifying the sample exe, it can't be strong-named, otherwise the signature is invalid.

Currently Samples.Probes.exe is strong-named, but it doesn't need to be; only its _dependencies_ are referenced by the integration project (which is strong named, because Datadog.Trace is strong named and we have InternalsVisibleTo).

This commit fixes the issue by moving the strong-naming to the dependencies only, meaning we can safely update Samples.Probes.exe

* Move set CorFlags value to a separate function

* Apply suggestions from code review

Co-authored-by: Zach Montoya <zach.montoya@datadoghq.com>
Co-authored-by: Lucas Pimentel-Ordyna <lucas.pimentel@datadoghq.com>

* Cache the _corFlags.exe value to avoid multiple look-ups

Co-authored-by: Zach Montoya <zach.montoya@datadoghq.com>
Co-authored-by: Lucas Pimentel-Ordyna <lucas.pimentel@datadoghq.com>