Proof of concept using Activity within Span/SpanContext
This is assuming that we take a new dependency on System.DiagnosticSource
for all necessary frameworks (built in for later versions of .NET)

Not in a functional state - copy/pasted some SpanContext ctors to
fix a few errors.

Could probably go through and refactor SpanContext to simply contain
the Parent SpanContext and TraceContext and just defer everything else
to the Activity.