Update Nuke build to latest (#4000)
* Update Nuke to the latest

```bash
dotnet tool install -g Nuke.GlobalTool
nuke :update
```

Fix build errors so that it compiles

* Fix build warnings

Mostly replacing Logger. with Serilog.Log

* nit: include all bootstrap files in proj

* Fix SetVariable

Clashes with existing SetVariable added in latest Nuke version

* Remove _build project from build config

* Exclude the build assets from Nuke.Common

For some reason, these cause MSBuild to fail with
OOM due to infinite recursion.

Narrowed it down to the new "convenience" including of
Directory.Build.props files in the Solution. Tried to
disable this by setting `$(NukeExcludeDirectoryBuild)=True`
but that _doesn't_ solve it for some reason.

As a workaround, have to exclude all the build assets.
They don't seem to be required for what we're doing,
but just in case (and to avoid confusion later), I copied
their contents to the Nuke project's Directory.Build.props
and Directory.Build.targets files, and commented out the
problematic section.

Note that there's a little hackery required in referencing
the Nuke.MSBuildTasks.targets file that comes with the
Nuke.Common NuGet package. Added a workaround, but that depends
on $(NuGetPackageRoot) being set, which doesn't seem to always
be the case (in CI on Windows specifically). As we're not using
them generally, I set `$(NukeTasksEnabled)=False` which skirts
around the issue

* Replace custom version of MSbuildSettings with built-in versions

* Just build everything

* Update build_in_docker.ps1

* Update build_in_docker.sh