Record initial product status versions
Not (currently) recording the values when they change in RCM because there are some lifetime and circular dep issues. Mostly I don't want to access static Tracer instances inside these singleton types, when they don't have them already.

We may be able to revisit this safely by refactoring LiveDebugger to be created in TracerManager (which theoretically should happen anyway) and/or updating RCM to pass an `ITelemetryController` instance when updating a subscription, but will leave that for a subsequent PR