Stop using concurrent dictionaries in metrics collection, and use direct indexing into an array instead.
Using a source generator, we pre-build an array of all the possible metric + tag combinations. We can then directly index into this array when recording metrics. To achieve this, we source-generate the method calls to write the metric.

The big advantage of this approach is that it replaces a concurrent dictionary access (which would be happening multiple times per-request) with direct access to an array, i.e. perf reasons.

Another advantage is that it documents exactly which tags are required for each metric, and their valid values. This was not possible with the previous design, and consequently there were errors.