Update Telemetry Metrics implementation for performance (#4251)
* Fix telemetry metric distribution type

* Stop using concurrent dictionaries in metrics collection, and use direct indexing into an array instead.

Using a source generator, we pre-build an array of all the possible metric + tag combinations. We can then directly index into this array when recording metrics. To achieve this, we source-generate the method calls to write the metric.

The big advantage of this approach is that it replaces a concurrent dictionary access (which would be happening multiple times per-request) with direct access to an array, i.e. perf reasons.

Another advantage is that it documents exactly which tags are required for each metric, and their valid values. This was not possible with the previous design, and consequently there were errors.

* Update metrics to use the new source generator design

* Update metric and Distribution DTOs to use string[] instead of List<string>

This means we can avoid recreating lists every time we send, as they're created once by the buffers

* Update the MetricsTelemetryCollector to use the new source-generated methods

* Add source-generator output

* Update usages of telemetry metrics API