Refactor `TelemetryControllerV2` to be a singleton (#4354)
* Fix not reusing previous values in telemetry controller

* Move tracking of AppStartedSent to DataAggregator

This is all information that needs to be tracked across telemetry restarts (i.e. DataAggregator needs to be a singleton) so makes sense to include it there I think

* Remove AppStartedSent from TelemetryControllerv2 as it's now unused

* Make telemetrycontrollerv2 a singleton and don't stop pushing on fatal errors

Necessary to handle situations in which telemetry is replaced e.g. in dynamic config.
As a side point, also decided to _not_ stop attempting to push on fatal errors, as this could be caused by an agent dropping off.

* Fix TelemetryControllerV2 initialization to avoid multiple copies

* Add missing dispose to unit test