Fix recording incorrect values in GetAs<>
Previously, when providing a T default value in configuration, we would call ToString() on the value to record it in telemetry. However, this may not be a reasonable value to record.

To work around that, the getDefaultValue() function now allows providing a string default value to record in telemetry. This ensures we can record a "reasonable" value for the default cases