[Tracer] Ensure the configuration of the trace context propagator is included in our telemetry (#4460)
* Update TracerSettings so that the default value of "tracecontext,Datadog" is always recorded in v1 and v2 telemetry. Also add integration tests and unit tests for this

* Alternatively, when the ActivityListener is enabled, only add the "tracecontext" propagator after telemetry is recorded so our v1 telemetry matches v2 telemetry, where the configuration value does not include the "tracecontext" value.

Note: This shouldn't come up very often since our default propagator already includes tracecontext, but it could occur in niche scenarios where the user has explicitly _not_ added tracecontext as a propagator.

* Updates to propagation telemetry recording (#4465)

* Revert "Alternatively, when the ActivityListener is enabled, only add the "tracecontext" propagator after telemetry is recorded so our v1 telemetry matches v2 telemetry, where the configuration value does not include the "tracecontext" value."

This reverts commit 79c7e713fd78be0d760db32114f78c1272476db4.

* Add "Unknown" configuration origin

* Fix recording incorrect values in GetAs<>

Previously, when providing a T default value in configuration, we would call ToString() on the value to record it in telemetry. However, this may not be a reasonable value to record.

To work around that, the getDefaultValue() function now allows providing a string default value to record in telemetry. This ensures we can record a "reasonable" value for the default cases

* Update propagation style injection and extraction to set the default value when extracting instead of overwriting later. Also make sure to record the values set in telemetry V2

---------

Co-authored-by: Andrew Lock <andrew.lock@datadoghq.com>