Move metrics aggregation loop to the metrics telemetry controller
By moving the metrics aggregator out of the telemetry controller, the aggregation shouldn't be affected if we have slow or mis-timed flushes

The downside to this is the aggregation and "serializing" must now be threadsafe, so opted for a simple lock.

The scheduler which was previously introduced to manage separate loops is overly complex currently, but will shortly have a new loop to add (with diagnostic logs) so kep it as-is for now.

, instead of being managed by the telemetry