Timeout after 3 seconds for Kafka and flush every 1 second
With earlier versions of the NuGet producing to an invalid topic would
return right away. However, in later versions the producer would
block for 30 seconds. This caused the tests to take around 1 minute each.

Specifying a timeout of 3 seconds maintains similar behavior as the
previous versions and keeps the test time down.