Assorted speed improvements to tests (#4479)
* Increase assertion range

Previously the 100ms assertion was failing occasionally.

* Abort response when not sending

Attempting to just speed up some slower tests. When we aren't sending
responses back in some tests (by design) we get blocked on the next call
to .GetContext() for 15 seconds.

By aborting it we don't send the response but won't get blocked on
the next call to .GetContext().

* Reduce wait time for IAST disabled tests

We aren't expecting any spans when it is disabled.

* Wait for 2 seconds for ProcessStartTests Disabled spans

* Set expected spans counts for ProcessStartTests for non-Windows

* Change expected spans to 3 for Linux for AgentMalfunctionTests

This should save a decent bit of time for tests.

* Timeout after 3 seconds for Kafka and flush every 1 second

With earlier versions of the NuGet producing to an invalid topic would
return right away. However, in later versions the producer would
block for 30 seconds. This caused the tests to take around 1 minute each.

Specifying a timeout of 3 seconds maintains similar behavior as the
previous versions and keeps the test time down.

* Document why we have a 30s Thread.Sleep

This is from Kevin's PR 3535

* Remove Task.Delays in RelationalDatabaseTestHarness

I don't understand why these are here and can't find anything really
in the git history for these.

This was causing each test to add 600ms of overhead which made many of
the database integration tests have an increased run time.

Removing them hasn't seemed to cause any additional flake.

* Apply suggestions from code review

Co-authored-by: Andrew Lock <andrew.lock@datadoghq.com>

* Add more PR suggestions

---------

Co-authored-by: Andrew Lock <andrew.lock@datadoghq.com>