[Tracing] Fix WebRequest bug with existing distributed tracing headers (#4770)
When we pre-emptively add distributed tracing headers in `WebRequest.BeginGetRequestStream`, `WebRequest.BeginGetResponse`, and `WebRequest.GetRequestStream`, add the headers to a cache to signal to the later integrations that we had last updated the headers and we are responsible for the distributed tracing headers in the WebRequest object.

When we check for the pre-emptively added distributed tracing headers in `WebRequest.EndGetResponse`, `WebRequest.GetResponse` and `WebRequest.GetResponseAsync`, first check a cache to know whether we added distributed tracing headers to the WebRequest object. If present, then continue to use the injected trace context. Otherwise, we create a new context.