Add configuration for sending (redacted) logs via telemetry
Initializes a Serilog sink that sends to the RedactedErrorLogSink. This in turn forwards to the RedactedErrorLogCollector.
Logs can be written immediately to the sink (which forwards to the collector) but they will not be sent until telemetry is enabled.

Note that this commit does not include the redaction or de-duping of logs functionality