Add redacted-error-log telemetry collector and sinks (#4833)
* Add telemetry logs payload

* Add RedactedErrorLogCollector

Add sending of logs to telemetry controller v2 based on schedule and queue-size
Separates logs into batches based on (very) rough calculation of maximum byte size. This always overestimates (significantly) but that's _probably_ ok?

* Add configuration for sending (redacted) logs via telemetry

Initializes a Serilog sink that sends to the RedactedErrorLogSink. This in turn forwards to the RedactedErrorLogCollector.
Logs can be written immediately to the sink (which forwards to the collector) but they will not be sent until telemetry is enabled.

Note that this commit does not include the redaction or de-duping of logs functionality

* Add integration tests for redacted error logs

* Handle the fact that the redacted error logs might not be enabled, so don't bother creating the collector and waiting for them

* Try to debug what's going on in replacing global tracer manager

* Remove unnecessary condition

* Handle case where telemetry is enabled by logs aren't