Don't send some error logs to telemetry (#4934)
* Downgrade "Error occured sending data to the agent"

It's an error, but one we can't do anything about. Warnings are written to the logs in the same way, so this shouldn't be a big deal hopefully?

Alternatively, we could apply a filter to the log telemetry logs

* Downgrade "The process does not have sufficient permissions to read performance counters" to warning

It's an error, but one we can't do anything about. Warnings are written to the logs in the same way, so this shouldn't be a big deal hopefully?

Alternatively, we could apply a filter to the log telemetry logs

* Revert "Downgrade "Error occured sending data to the agent""

This reverts commit bef6b345d509397e32efa4b11c17403bd52e9a32.

* Revert "Downgrade "The process does not have sufficient permissions to read performance counters" to warning"

This reverts commit 677fa4e148909b95e5d1a17504fc60e5aff84d7a.

* Exclude specific logs that we can't do anything about

* Add unit test for not sending the message