Update snapshots for DD_TRACE_DELAY_WCF_INSTRUMENTATION_ENABLED=true
Notably, this results in two (yes two) error spans. This is because we make the Begin/End async callbacks for the WcfSample/ICalculator/ServerAsyncAdd endpoint throw an exception. One time we throw an exception during the Begin method and one time we throw an exception during the End method. Previously, we were only capturing one exception instead of two.