Add support for sync Read calls.
It's only used in the Mock agent currently, so should be "safe" as long as we make sure to avoid any sync stream interactions in Datadog.Trace.

It's not ideal, but refactoring to handle both async-and-sync paths doesn't seem worth it.

Also removed the Debug.Fail() because the mock form parser calls Read on a stream that has already returned 0