Create a new NLog configuration when necessary
This is because in v5.3+ of NLog a call was
changed to start requiring that the
LoggingConfiguration object within NLog was not
null to create all of the various LoggingRules.

With this change, our instrumentation for direct
log shipping wasn't working if there was no
NLog configuration.

This fixes then.