Fallback to operationName instead of empty string for ElasticSearch ResourceName
In some logs I saw that the trace agent was needing to fix many
ElasticSearch spans where the ResourceName was empty - it
would populate the ResourceName with the operationName.

It seems like we could just do that in the Tracer instead.

I'm unsure exactly how the ElasticSearch RequestParameters are null
here to cause this issue though.