For the OpenTelemetry configuration mapping, there's a different configuration logic we want to introduce: If any of the Datadog configurations were set, even if it's invalid (like a value that should be parsed as an int but is not an int), we want to ignore the OpenTelemetry configuration.
This commit adds a new ITelemeteredConfigurationSource.IsPresent method to query a configuration source if a key is present, and adds it to the OpenTelemetry configuration mapping logic.

This also updates some test cases where the OpenTelemetry configuration was previously being consulted when Datadog configurations were present but invalid. Now the OpenTelemetry configuration is ignored in that scenario, as desired.