Only reset Activity.Id for W3C format
If we attempt to reset Activity.Id for non-W3C
formats we'll create a null Id -> it doesn't
get recreated like it does for the W3C format.

This leads to the Activity.Id being null -
which is something that we noticed recently in
Error Tracking but were unsure how it happened.

This was being caused by a start time of an
Activity being modified after being started to
an earlier time. This is fine for W3C format
Activity objects but not for non-W3C as all
they have is the .Id value which would be set to
null.