Config refactor - distinguish between "not present" and "not valid" (#5713)
## Summary of changes

- Changes `ITelemetredConfigurationSource` methods to return
`ConfigurationResult<T>` instead of `ConfigurationResult<T>?`
- Add additional properties to `ConfigurationResult<>` to distinguish
between "key not found" and "value not parsable"

## Reason for change

In https://github.com/DataDog/dd-trace-dotnet/pull/5661, we added
support for mapping standard OTel configuration to DD equivalents.
Unfortunately, doing this correctly required distinguishing between "key
not found" and "value not parsable". As a workaround, an addition method
`IsPresent(key)` was added, but this essentially duplicates a lot of the
work we're already doing when loading keys "normally".

The changes in this PR are a first step to simplifying and unifying the
support for Otel

## Implementation details

- Change all `ITelemeteredConfigurationSource.Get*` methods to return
`ConfigurationResult` instead of a potentially null value. Update all
sources as appropriate.
- Update `ConfigurationResult`
  - Add `bool IsPresent()`
- Add `LoadResult` enum which can be one of `Valid`, `NotFound`,
`ParsingError`, or `ValidationFailure `. Previously we effectively had a
bool of just `Valid` or `ValidationFailure`, and "is null" meant
`NotFound` _or_ `ParsingError`
  - Add helper method `ShouldFallBack`
- Refactor `ConfigurationBuilder`
- Extract common "cascading fallbacks" code to private `GetResult`
methods. By extracting static lambdas we can reuse a lot of the logic
for multiple types
  - Refactor code to use the new properties where possible. 
- `ITelemeteredConfigurationSource.IsPresent` is still used in Otel
paths currently - removed in a subsequent PR

## Test coverage

This is just a refactor, so all covered by existing tests

## Other details

This is part of a big stack of config refactoring PRs:

- https://github.com/DataDog/dd-trace-dotnet/pull/5713 (this PR)
- https://github.com/DataDog/dd-trace-dotnet/pull/5714
- https://github.com/DataDog/dd-trace-dotnet/pull/5715
- https://github.com/DataDog/dd-trace-dotnet/pull/5716
- https://github.com/DataDog/dd-trace-dotnet/pull/5717

<!-- ⚠️ Note: where possible, please obtain 2 approvals prior to
merging. Unless CODEOWNERS specifies otherwise, for external teams it is
typically best to have one review from a team member, and one review
from apm-dotnet. Trivial changes do not require 2 reviews. -->