Revert "Add workflow monitor to all workflows (#5733)" (#5741)
## Summary of changes

Reverts the workflow monitor

## Reason for change

It appears to absolutely _cripple_ the GitHub action network speed.
Actions have been failing on master ever since we merged this PR, but
here are some examples:

Checkout stage
- Before: 9s
- After: 3min 30s

Install .NET stage 
- Before: 7s
- After: 7min

Download Azure Build assets
- Before: 40s (it's _big_)
- After: timeout after 140s

## Implementation details

Revert the monitor, we'll have to YOLO the permissions changes with what
we have

## Test coverage

I've tested this is the cause by disabling it using the variable. But if
it's disabled, there's kind of no point in including it in the
workflows, so may as well tear it out

## Other details

grmbl grmbl, gonna have to just YOLO our permission changes in that case

<!-- ⚠️ Note: where possible, please obtain 2 approvals prior to
merging. Unless CODEOWNERS specifies otherwise, for external teams it is
typically best to have one review from a team member, and one review
from apm-dotnet. Trivial changes do not require 2 reviews. -->