[v3] Annotate duck type proxies with type information (#5742)
- Adds the ability to annotate duck types with type + assembly
information
- Annotates the types used with Datadog.Trace.Manual
- Adds unit tests to make sure we don't break that duck typing contract
- Fix minor bugs in instrumentation definitions

We need to be very careful we don't break duck types for our
Datadog.Trace.Manual API when evolving Datadog.Trace. We have
integration tests for this already, so this is just an extra layer of
defense to make sure we don't change things without realizing the
consequences.

Additionally, the proxy type annotations are required as part of the
Native AOT work, so the attributes are based on work already started by
@kevingosse in a branch:
https://github.com/DataDog/dd-trace-dotnet/commit/a41cf637efd334c722b1c6e22f89b320609b3ebf

- Add `[DuckType]` attribute, added to proxies to indicate the types
they are designed to target
- Add overload to `[DuckCopy]` for same purpose (I added this because
it's what Kevin did, but maybe we should just stick to `[DuckType]`, and
have you add both?)
- Add attributes to all proxies in Datadog.Trace that proxy something in
Datadog.Trace.Manual
- Add attributes to all types in _Datadog.Trace.Manual_ that proxy
something in _Datadog.Trace_!
- For example, the Datadog.Trace instrumentation creates a
`Datadog.Trace.Scope` and then duck types that to a
_Datadog.Trace.Manual_ `IScope`
- Things get a bit messy when we're "reusing" types by linking into both
projects, e.g. `IScope`. Made the interfaces partial to work around it
for now, but we can probably simplify this down the line

Adds extra coverage

One oddity, is that I couldn't seem to `DuckCast` `IScope` to `IScope` 🤔
The duck typing complains about the missing (inherited) member `void
Dispose()`. Ignored it for now, as in practice we're always (currently)
duck typing a `Scope` which _does_ work, but we may need to look into it
again later

<!-- ⚠️ Note: where possible, please obtain 2 approvals prior to
merging. Unless CODEOWNERS specifies otherwise, for external teams it is
typically best to have one review from a team member, and one review
from apm-dotnet. Trivial changes do not require 2 reviews. -->