Re-order integrations folder in CODEOWNERS (#5762)
## Summary of changes

Re-order the ownership of the `integrations` folder.

## Reason for change

This was adding the tracing and idm team to reviewers of PR that were
not needed.
## Implementation details

## Test coverage

## Other details
<!-- Fixes #{issue} -->

<!-- ⚠️ Note: where possible, please obtain 2 approvals prior to
merging. Unless CODEOWNERS specifies otherwise, for external teams it is
typically best to have one review from a team member, and one review
from apm-dotnet. Trivial changes do not require 2 reviews. -->