Add attribute for skipping tests in CI without using `[Fact(Skip = "")]` (#5756)
## Summary of changes

Adds a new trait, `SkipInCI` that can be used to skip tests in CI

## Reason for change

We currently use the built-in `[Fact(Skip="..."])` or
`[Theory(Skip="..."])` to skip tests that we don't want to run in CI.
The downside to this is if you want to run a test locally, you have to
explicitly change the code to remove the `Skip`, run the test, and
remember to add it back later.

Instead, we can change the default `Filter` used by Nuke for each stage
to be `SkipInCI!=True`. That way the tests are still excluded from CI
(or if you run the "default" test suite locally), but you can still
easily run the test without making code modifications if you want to run
it directly from VS/Rider, or using `nuke RunWindowsIntegrationTests
-Filter SomeSkippedTest` (for example).

## Implementation details

Replaced existing usages of 
```csharp
[Fact(Skip="Some reason"])
```

with 

```
[Trait("SkipInCI", "True")] // Some reason
```

and changed the default filters used by Nuke to exclude tests with this
trait

## Test coverage

This PR is the test - if it passes, we're good.

## Other details

It would be nice to have the skip reason as the _value_ of the trait,
but the [dotnet test
syntax](https://learn.microsoft.com/en-us/dotnet/core/testing/selective-unit-tests?pivots=xunit)
doesn't make it easy to select just based on the key, it's Key=Value, so
we're stuck with comments.

<!-- ⚠️ Note: where possible, please obtain 2 approvals prior to
merging. Unless CODEOWNERS specifies otherwise, for external teams it is
typically best to have one review from a team member, and one review
from apm-dotnet. Trivial changes do not require 2 reviews. -->