Revert native changes that moved the definitions to the native side (#5768)
## Summary of changes

Reverts recent change that moved definitions to the native side

## Reason for change

We've seen crashes on the native side in smoke tests etc that we haven't
got to the bottom of yet. Out of caution, choosing to revert these
changes to investigate further

## Implementation details

Reverted the following PRs: 
- https://github.com/DataDog/dd-trace-dotnet/pull/5757 (introduced to
fix a bug in 5592)
- https://github.com/DataDog/dd-trace-dotnet/pull/5592

## Test coverage

This is the main test, will run a full installer test suite on this PR
to confirm we don't see the issues immediately. Will require further
monitoring to look for other issues

## Other details

A less drastic version of 
- https://github.com/DataDog/dd-trace-dotnet/pull/5767

as we expect the issue is in the moved definitions.

<!-- ⚠️ Note: where possible, please obtain 2 approvals prior to
merging. Unless CODEOWNERS specifies otherwise, for external teams it is
typically best to have one review from a team member, and one review
from apm-dotnet. Trivial changes do not require 2 reviews. -->