[v3] Package musl assets in linux glibc installers (#5595)
## Summary of changes

- Adds the musl assets to the Linux glibc installers
- Add smoke tests for using the glibc installer on alpine

## Reason for change

To simplify onboarding, we now include the musl assets in the glibc
installer packages. Instead of having to choose which package to
install, there's a single "correct" choice.

For now, we have left the `-musl.tar.gz` package with only the `-musl`
artifacts (i.e. we didn't package the glibc assets there). The reasoning
was that customers have already _explicitly_ selected musl at this
point, so the glibc assets are very unlikely to be used. Alpine
customers are also likely more size-conscious. We could potentially
consider deprecating the artifact in a future major release, but don't
plan to currently.

## Implementation details

Package the musl assets in the glibc installers (in the `linux-musl-x64`
path). Note that the universal binaries (ld_preload, native loader) are
identical, so we hardlink those in to ensure we don't increase package
size too much.

## Test coverage

Added smoke tests to confirm that we can install and run with the
universal (glibc) package when running on alpine.

Checking the size of the assets:

- Before: ~20MB compressed (.tar.gz etc size), ~60MB uncompressed (on
disk)
- After: ~30MB compressed (.tar.gz etc size), ~80MB uncompressed (on
disk)

## Other details

Currently the dd-dotnet.sh script directs customers to install the
`-musl` package when running on Alpine (and it detects you've installed
the wrong package). Should we point to the universal package instead?
Bit of a moot point, as we shouldn't _ever_ hit that branch now (if all
is working well)

Requires:
- https://github.com/DataDog/dd-trace-dotnet/pull/5550
- https://github.com/DataDog/dd-trace-dotnet/pull/5593