[Dynamic Instrumentation] Fixed instrumentation error (InvalidProgramException) related retargeting of EH clauses (#5774)
## Summary of changes
Fixed `InvalidProgramException` thrown due to malformed bytecode born
from DI Line Probe instrumentation, discovered by Exploration Testing of
Method + Line Probes.

## Reason for change
When a customer chooses a line for instrumentation through DI, a bunch
of bytecode is being placed in front of the chosen line, to capture the
method state.
When the original line was a target of an exception handling clauses
(e.g the line was the beginning of a catch/finally blocks, the line was
the end of a try block, etc) - in those cases, we retarget the exception
handling pointers to point to the new bytecode being placed. The code
that handled it missed two assignment: TryEnd and HandlerEnd. Meaning,
when the target line was an end of finally/catch block or at the end of
a try block, we missed fixing the pointers, resulting in
`InvalidProgramException`.

## Implementation details
Correctly assinging `TryEnd` & `HandlerEnd` EH clauses.

## Test coverage
Four new tests were added, that previously were broken with
`InvalidProgramException`:
- `TryFinallyTest`.
- `AsyncTryFinallyTest`
- `TryCatchTest`
- `AsyncTryCatchTest`