Integration test for Oracle (#5607)
## Summary of changes

There was an Oracle Sample app, but it was not used in integration
tests.
I added a minimal test that at least allows manual testing, and added
the required docker things.

## Implementation details

I also merged the two Oracle sample projects (core and non-core) into
one, using csproj conditions.

The Oracle image does not work (at all) on mac in docker desktop. I used
[colima](https://github.com/abiosoft/colima) to test it, which worked
great (but I don't think it's something we want to use in the CI, so I
marked the test as non-arm).

The image for the oracle DB is quite big, so we'll have to keep an eye
on the CI perf with this on.