Correctly set `activityKey` when we change the Trace ID (#5771)
## Summary of changes

This delays the computation of `activityKey` (used to map an `Activity`
to a Datadog `Scope`) by it's `TraceID` + `SpanID` to be _after_ we
could potentially modify those values.

## Reason for change

We do modify the `TraceId` of an `Activity`. The previous way would
cause `MISSING SCOPE` logs as we'd use the outdated Trace Id when
searching for the mapping.

## Implementation details

- Move setting of `activityKey` later
- Update `activityTraceId` when we change the Activity's TraceId

## Test coverage

- Had to update the `OpenTelemetryTests` as they weren't correct

## Other details
<!-- Fixes #{issue} -->

<!-- ⚠️ Note: where possible, please obtain 2 approvals prior to
merging. Unless CODEOWNERS specifies otherwise, for external teams it is
typically best to have one review from a team member, and one review
from apm-dotnet. Trivial changes do not require 2 reviews. -->