Change unknown_service to DefaultServiceName (#5671)
## Summary of changes

This looks for an `unknown_service` in `service.name` for
`Activity`/`OpenTelemetry` spans and if detected will replace that with
the value of `DefaultServiceName`.

## Reason for change

OpenTelemetry will use `unknown_service:AssemblyName` as the default for
`service.name` if there isn't a value for it.
A workaround to this is to have `OTEL_SERVICE_NAME` set to the same
value as `DD_ENV`, this avoids needing to do that workaround.

## Implementation details

Avoid `unknown_service`
https://opentelemetry.io/docs/specs/semconv/resource/#service

## Test coverage

Updated the OpenTelemetrySdk tests to replicate this.

## Other details
<!-- Fixes #{issue} -->

Fixes #5650

<!-- ⚠️ Note: where possible, please obtain 2 approvals prior to
merging. Unless CODEOWNERS specifies otherwise, for external teams it is
typically best to have one review from a team member, and one review
from apm-dotnet. Trivial changes do not require 2 reviews. -->